requirements.txt
streamlit
pillow
google-generativeai
python-dotenv
pillow-heif